ビルドエラー修正

+++ rg-etc1/src/rg_etc1.cpp
@@ -2104,7 +2104,7 @@ namespace rg_etc1
          {
             if (block_inten[0] > m_pSorted_luma[n - 1])
             {
-               const uint min_error = labs(block_inten[0] - m_pSorted_luma[n - 1]);
+               const uint min_error = static_cast<uint>(labs(static_cast<int>(block_inten[0]) - static_cast<int>(m_pSorted_luma[n - 1])));
                if (min_error >= trial_solution.m_error)
                   continue;
             }
@@ -2118,7 +2118,7 @@ namespace rg_etc1
          {
             if (m_pSorted_luma[0] > block_inten[3])
             {
-               const uint min_error = labs(m_pSorted_luma[0] - block_inten[3]);
+               const uint min_error = static_cast<uint>(labs(static_cast<int>(m_pSorted_luma[0]) - static_cast<int>(block_inten[3])));
                if (min_error >= trial_solution.m_error)
                   continue;
             }
